sync läuft, probleme mit delete all -> nochmal testen

+++ app/src/main/java/org/dieschnittstelle/mobile/android/skeleton/model/RetrofitToDoItemCRUDOperationsImpl.java
@@ -108,7 +108,11 @@ public class RetrofitToDoItemCRUDOperationsImpl implements IToDoItemCRUDOperatio
 
     @Override
     public boolean deleteAllTodoItems(boolean remote) {
-        return deleteAllTodos();
+        if (remote){
+            return deleteAllTodos();
+        } else {
+            return false;
+        }
     }
 
     @Override
